requirements.txt
astroid==2.5.1
autopep8==1.5.6
blurhash-python==1.1.0
cffi==1.14.5
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.7.0
pycparser==2.20
pylint==2.7.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
